razor pay satrted

+++ models/category.go
@@ -15,8 +15,8 @@ type Products struct {
 	gorm.Model
 	Product_Name   string `json:"product_name" binding:"required"`
 	Product_Colour string `json:"product_colour" binding:"required"`
-	Product_Size   int16  `json:"product_size" binding:"required"`
+	Product_Size   int  `json:"product_size" binding:"required"`
 	Product_Brand  string `json:"product_brand" binding:"required"`
-	Product_Price  int16
-	Stock          int16
+	Product_Price  int
+	Stock          int
 }
